==> code/02_parse_and_segment_audio.ipynb <==
Installing background package and importing libraries that will be used

# !pip install pydub

import os
import pydub
import pandas as pd
from scipy.io import wavfile
from pydub import AudioSegment
from pydub.silence import split_on_silence, detect_silence

pydub.AudioSegment.converter = "/Users/allanshomefolder/anaconda3/bin/ffmpeg" #Insert path here r"C:/Users/kwama/Downloads/ffmpeg/ffmpeg/bin/ffmpeg.exe"
pydub.AudioSegment.ffmpeg = "/Users/allanshomefolder/anaconda3/bin/ffmpeg" #Insert path here r"C:/Users/kwama/Downloads/ffmpeg/ffmpeg/bin/ffmpeg.exe"r"C:/Users/kwama/Downloads/ffmpeg/ffmpeg/bin/ffmpeg.exe"
pydub.AudioSegment.ffprobe = "/Users/allanshomefolder/anaconda3/bin/ffprobe" #Insert path here r"C:/Users/kwama/Downloads/ffmpeg/ffmpeg/bin/ffmpeg.exe"r"C:/Users/kwama/Downloads/ffmpeg/ffmpeg/bin/ffprobe.exe"

We used the pydub Python package to parse and segment our mp3 files. Pydub accepts mp3 files and has methods that detect silence, detect audio, and split the mp3 files based on silence. We utilized the last method to pull samples of audio from our data. There are two parameters we passed in as well -- min_silence_len and silence_thresh. These determine how long a pause needs to be in milliseconds before it’s registered as silence, and how loud something needs to be in dbFS before its registered as audio. We played with a few min_silence_len and silence_thresh values, but settled on 2.5 seconds and -16dbFS for our dataset. 

sound_file = AudioSegment.from_file("../BArT/testing/mp3_downloads/18812-20191027-0001.mp3", format="mp3")
channel_count = sound_file.channels    
sample_width = sound_file.sample_width 
duration_in_sec = len(sound_file) / 1000 
sample_rate = sound_file.frame_rate

print ("sample_width=", sample_width)
print ("channel_count=", channel_count)
print ("duration_in_sec=", duration_in_sec)
print ("frame_rate=", sample_rate)

file = os.listdir('../audio_data/audio_files/mp3_files/')[1]
filename = '../audio_data/audio_files/mp3_files/' + file
audio = AudioSegment.from_file(filename, format="mp3")
audio_chunks = split_on_silence(audio, 
                              min_silence_len = 2_500,  #milliseconds          
                              silence_thresh = -16)

# mkdir f'wav samples for {file}' ??

Converts audio samples that are between 1 second and 60 seconds long from mp3 to wav format. Google Speech to Text API accepts wav files, not mp3 files, and only allows for a total of 60 minutes of API use before you have to start paying for their services. We took that into account when deciding how long to make our samples.

#Converst and adds audio samples that are between 1 and 60 seconds long 
for i, chunk in enumerate(audio_chunks):
    if (len(chunk) > 1000) and (len(chunk) < 60_000):
        fn = f"./audio_data/audio_files/wav_files/{file.split('.')[0]}_{i}.wav"
        chunk.export(fn, format="wav")


>Code adopted from [NY General Assembly DSI radio-to-location repository](https://github.com/mchbmn/radio-to-location)



---

## J's attempt to clean up the audio

----

import requests as req
